requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.3.4
billiard==3.6.4.0
celery==4.4.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.2.3
django-factory-boy==1.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
factory-boy==3.2.0
Faker==8.2.1
idna==2.10
kombu==4.6.11
prompt-toolkit==3.0.18
PyJWT==2.1.0
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.4
vine==1.3.0
wcwidth==0.2.5
